+++ NEWS.md
@@ -1,3 +1,11 @@
+# cthist 1.3.0
+
+* Added "results posted" (indicates whether study results have been
+  posted)
+* Added "references" to `clinicaltrials_gov_version()` (citations,
+  links and available IPD/Information) and `drks_de_version()` (Trial
+  Publications, Results and other Documents)
+
 # cthist 1.2.1
 
 * Implemented checks in `clinicaltrials_gov_download()` and